[FEAT] Save/Load 기능 템플릿 메소드 패턴으로 구현
