--- DateWithKing/Assets/Scripts/Screen/SaveScreen.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e8f00b1bbbec1164aa39a1ba3171e580
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
