--- DateWithKing/Assets/Scripts/Screen/SaveScreen/SaveScreenView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bf2ea7a1d9c98f24aba9b565c549ce9f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
